NextJS 10->11 upgrade
After upgrade, it only works with Webpack5 set to false. NextJS v12 will not be compatible with custom webpack config that is currently being used.  Some fix needs to be done. Test of removing the webpack custom config on next.config.js does not work, it will crash on _document.js. Not sure if that crash is related to material-ui, since Material UI (changed to MUI) supports React 17 in a different namespace @mui. Old @material-ui seems to work with React 17, but taking way the webpack config will just crash.  Lots of things have dependency on the old @material-ui namespace, so this is like a nightmare situation.